RefreshContinuation分割でfile_length解消
