+++ YoutubeFeeder/Features/FeedCache/FeedCacheCoordinator+RefreshContinuation.swift
@@ -1,17 +1,37 @@
 import Foundation
 
-extension FeedCacheCoordinator {
+@MainActor
+final class FeedCacheCoordinatorRefreshContinuation {
+    unowned let coordinator: FeedCacheCoordinator
+    private lazy var support = FeedCacheCoordinatorRefreshContinuationSupport(coordinator: coordinator)
+
+    init(coordinator: FeedCacheCoordinator) {
+        self.coordinator = coordinator
+    }
+
+    func beginManualRefreshProgress(totalChannels: Int) {
+        support.beginManualRefreshProgress(totalChannels: totalChannels)
+    }
+
+    func updateManualRefreshActiveCalls(completed: Int, totalChannels: Int, activeCalls: Int) {
+        support.updateManualRefreshActiveCalls(completed: completed, totalChannels: totalChannels, activeCalls: activeCalls)
+    }
+
+    func finishManualRefreshProgress() {
+        support.finishManualRefreshProgress()
+    }
+
     func performMockRefresh() async {
-        let totalChannels = max(progress.totalChannels, maintenanceItems.count)
-        refreshProgress = CacheRefreshProgress(
+        let totalChannels = max(coordinator.progress.totalChannels, coordinator.maintenanceItems.count)
+        coordinator.refreshProgress = CacheRefreshProgress(
             isRefreshing: true,
             checkStage: RefreshStageProgress(title: "フィード更新確認", completed: totalChannels, total: totalChannels, activeCalls: 0, callsPerSecond: 3),
             fetchStage: .idle(title: "更新チャンネル取得", callsPerSecond: 0),
             thumbnailStage: .idle(title: "サムネイル取得", callsPerSecond: 0)
         )
-        await refreshUI(currentChannelID: nil, isRunning: false, lastError: progress.lastError)
-        refreshProgress.isRefreshing = false
-        _ = await performConsistencyMaintenanceIfNeeded(force: false)
+        await refreshUI(currentChannelID: nil, isRunning: false, lastError: coordinator.progress.lastError)
+        coordinator.refreshProgress.isRefreshing = false
+        _ = await coordinator.performConsistencyMaintenanceIfNeeded(force: false)
     }
 
     func performMockChannelRefresh(channelID: String) async {
@@ -20,7 +40,7 @@ extension FeedCacheCoordinator {
             detail: "モック経路でチャンネル更新を処理",
             metadata: ["channelID": channelID]
         )
-        refreshProgress = CacheRefreshProgress(
+        coordinator.refreshProgress = CacheRefreshProgress(
             isRefreshing: true,
             checkStage: RefreshStageProgress(title: "チャンネル更新", completed: 0, total: 1, activeCalls: 1, callsPerSecond: 1),
             fetchStage: .idle(title: "更新チャンネル取得", callsPerSecond: 0),
@@ -29,14 +49,14 @@ extension FeedCacheCoordinator {
         await refreshUI(
             currentChannelID: channelID,
             isRunning: false,
-            lastError: progress.lastError
+            lastError: coordinator.progress.lastError
         )
         try? await Task.sleep(nanoseconds: 2_000_000_000)
-        refreshProgress = .idle
+        coordinator.refreshProgress = .idle
     }
 
     func performManualChannelRefresh(channelID: String) async {
-        let isRegisteredChannel = channels.contains(channelID)
+        let isRegisteredChannel = coordinator.channels.contains(channelID)
         RuntimeDiagnostics.shared.record(
             "channel_manual_refresh_fetch_started",
             detail: "フィード取得を開始",
@@ -45,14 +65,14 @@ extension FeedCacheCoordinator {
                 "registered": isRegisteredChannel ? "true" : "false"
             ]
         )
-        refreshProgress = CacheRefreshProgress(
+        coordinator.refreshProgress = CacheRefreshProgress(
             isRefreshing: true,
             checkStage: RefreshStageProgress(title: "チャンネル更新", completed: 0, total: 1, activeCalls: 1, callsPerSecond: 1),
             fetchStage: RefreshStageProgress(title: "更新チャンネル取得", completed: 0, total: 1, activeCalls: 0, callsPerSecond: 1),
             thumbnailStage: RefreshStageProgress(title: "サムネイル取得", completed: 0, total: 0, activeCalls: 0, callsPerSecond: 1)
         )
 
-        let result = await channelSyncService.performForcedRefresh(channelID: channelID)
+        let result = await coordinator.channelSyncService.performForcedRefresh(channelID: channelID)
         RuntimeDiagnostics.shared.record(
             "channel_manual_refresh_fetch_finished",
             detail: result.errorMessage == nil ? "フィード取得成功" : "フィード取得失敗",
@@ -63,12 +83,12 @@ extension FeedCacheCoordinator {
             ]
         )
         if result.errorMessage == nil {
-            await applyForcedRefreshSuccess(result, channelID: channelID)
+            await support.applyForcedRefreshSuccess(result, channelID: channelID)
         } else {
-            applyForcedRefreshFailure()
+            support.applyForcedRefreshFailure()
         }
 
-        let cleanup = isRegisteredChannel ? await performConsistencyMaintenanceIfNeeded(force: false) : nil
+        let cleanup = isRegisteredChannel ? await coordinator.performConsistencyMaintenanceIfNeeded(force: false) : nil
         RuntimeDiagnostics.shared.record(
             "channel_manual_refresh_maintenance_finished",
             detail: isRegisteredChannel ? "整合性メンテナンスを完了" : "未登録チャンネルのため整合性メンテナンスを省略",
@@ -83,16 +103,7 @@ extension FeedCacheCoordinator {
             isRunning: false,
             lastError: result.errorMessage
         )
-        refreshProgress = .idle
-    }
-
-    func beginManualRefreshProgress(totalChannels: Int) {
-        refreshProgress = CacheRefreshProgress(
-            isRefreshing: true,
-            checkStage: RefreshStageProgress(title: "フィード更新確認", completed: 0, total: totalChannels, activeCalls: 0, callsPerSecond: 3),
-            fetchStage: .idle(title: "更新チャンネル取得", callsPerSecond: 0),
-            thumbnailStage: .idle(title: "サムネイル取得", callsPerSecond: 0)
-        )
+        coordinator.refreshProgress = .idle
     }
 
     func runManualRefreshChannels(
@@ -104,10 +115,10 @@ extension FeedCacheCoordinator {
         var cycleResult = FeedRefreshCycleResult()
         let progressInterval = 50
         let cycleStartedAt = Date()
-        updateManualRefreshActiveCalls(completed: 0, totalChannels: sortedChannels.count, activeCalls: sortedChannels.isEmpty ? 0 : 1)
+        support.updateManualRefreshActiveCalls(completed: 0, totalChannels: sortedChannels.count, activeCalls: sortedChannels.isEmpty ? 0 : 1)
 
         for (index, channelID) in sortedChannels.enumerated() {
-            let result = await self.processChannel(
+            let result = await coordinator.processChannel(
                 channelID,
                 states: states,
                 forceNetworkFetch: forceNetworkFetch
@@ -132,7 +143,7 @@ extension FeedCacheCoordinator {
                     ]
                 )
             }
-            updateManualRefreshActiveCalls(
+            support.updateManualRefreshActiveCalls(
                 completed: completed,
                 totalChannels: sortedChannels.count,
                 activeCalls: remaining > 0 ? 1 : 0
@@ -142,103 +153,6 @@ extension FeedCacheCoordinator {
         return cycleResult
     }
 
-    func updateManualRefreshActiveCalls(completed: Int, totalChannels: Int, activeCalls: Int) {
-        refreshProgress.checkStage = RefreshStageProgress(
-            title: refreshProgress.checkStage.title,
-            completed: completed,
-            total: totalChannels,
-            activeCalls: activeCalls,
-            callsPerSecond: refreshProgress.checkStage.callsPerSecond
-        )
-    }
-
-    func finishManualRefreshProgress() {
-        refreshProgress = CacheRefreshProgress(
-            isRefreshing: false,
-            checkStage: completedStage(refreshProgress.checkStage),
-            fetchStage: completedStage(refreshProgress.fetchStage),
-            thumbnailStage: completedStage(refreshProgress.thumbnailStage)
-        )
-    }
-
-    func applyForcedRefreshSuccess(_ result: FeedChannelForcedRefreshResult, channelID: String) async {
-        refreshProgress.fetchStage = RefreshStageProgress(
-            title: refreshProgress.fetchStage.title,
-            completed: 0,
-            total: 1,
-            activeCalls: 1,
-            callsPerSecond: refreshProgress.fetchStage.callsPerSecond
-        )
-        refreshProgress.checkStage = completedStage(refreshProgress.checkStage, total: 1)
-        refreshProgress.fetchStage = completedStage(refreshProgress.fetchStage, total: 1)
-
-        let thumbnailTargets = result.uncachedVideos.filter { $0.thumbnailURL != nil }
-        refreshProgress.thumbnailStage = RefreshStageProgress(
-            title: refreshProgress.thumbnailStage.title,
-            completed: 0,
-            total: thumbnailTargets.count,
-            activeCalls: 0,
-            callsPerSecond: refreshProgress.thumbnailStage.callsPerSecond
-        )
-
-        for (index, video) in thumbnailTargets.enumerated() {
-            await cacheForcedRefreshThumbnail(video, channelID: channelID, index: index, total: thumbnailTargets.count)
-        }
-    }
-
-    func cacheForcedRefreshThumbnail(_ video: YouTubeVideo, channelID: String, index: Int, total: Int) async {
-        RuntimeDiagnostics.shared.record(
-            "channel_manual_refresh_thumbnail_started",
-            detail: "サムネイル取得を開始",
-            metadata: [
-                "channelID": channelID,
-                "videoID": video.id,
-                "index": String(index + 1),
-                "total": String(total)
-            ]
-        )
-        refreshProgress.thumbnailStage = RefreshStageProgress(
-            title: refreshProgress.thumbnailStage.title,
-            completed: index,
-            total: total,
-            activeCalls: 1,
-            callsPerSecond: refreshProgress.thumbnailStage.callsPerSecond
-        )
-        await writeService.cacheThumbnail(for: video)
-        RuntimeDiagnostics.shared.record(
-            "channel_manual_refresh_thumbnail_finished",
-            detail: "サムネイル取得を完了",
-            metadata: [
-                "channelID": channelID,
-                "videoID": video.id,
-                "index": String(index + 1)
-            ]
-        )
-        refreshProgress.thumbnailStage = RefreshStageProgress(
-            title: refreshProgress.thumbnailStage.title,
-            completed: index + 1,
-            total: total,
-            activeCalls: 0,
-            callsPerSecond: refreshProgress.thumbnailStage.callsPerSecond
-        )
-    }
-
-    func applyForcedRefreshFailure() {
-        refreshProgress.checkStage = completedStage(refreshProgress.checkStage, total: 1)
-        refreshProgress.fetchStage = completedStage(refreshProgress.fetchStage, total: 1)
-    }
-
-    func completedStage(_ stage: RefreshStageProgress, total: Int? = nil, completed: Int? = nil) -> RefreshStageProgress {
-        let resolvedTotal = total ?? stage.total
-        return RefreshStageProgress(
-            title: stage.title,
-            completed: completed ?? resolvedTotal,
-            total: resolvedTotal,
-            activeCalls: 0,
-            callsPerSecond: stage.callsPerSecond
-        )
-    }
-
     func refreshUI(
         currentChannelID: String?,
         isRunning: Bool,
@@ -246,14 +160,14 @@ extension FeedCacheCoordinator {
         includesVideos: Bool = true
     ) async {
         let startedAt = Date()
-        logRefreshUIStart(
+        support.logRefreshUIStart(
             currentChannelID: currentChannelID,
             includesVideos: includesVideos
         )
-        let refreshState = await readService.loadRefreshState(.init(
-            channels: channels,
-            freshnessInterval: freshnessInterval,
-            videoQuery: videoQuery,
+        let refreshState = await coordinator.readService.loadRefreshState(.init(
+            channels: coordinator.channels,
+            freshnessInterval: coordinator.freshnessInterval,
+            videoQuery: coordinator.videoQuery,
             currentChannelID: currentChannelID,
             isRunning: isRunning,
             lastError: lastError,
@@ -263,24 +177,24 @@ extension FeedCacheCoordinator {
         let nextProgress = refreshState.progress
         let nextMaintenanceItems = refreshState.maintenanceItems
 
-        applyRefreshUIState(
+        support.applyRefreshUIState(
             progress: nextProgress,
             maintenanceItems: nextMaintenanceItems,
             currentChannelID: currentChannelID,
             includesVideos: includesVideos
         )
-        await refreshHomeSystemStatus(snapshot: refreshState.snapshot, currentProgress: nextProgress)
+        await coordinator.refreshHomeSystemStatus(snapshot: refreshState.snapshot, currentProgress: nextProgress)
         let homeStatusUpdatedAt = Date()
 
-        await writeService.persistBootstrap(progress: progress, maintenanceItems: maintenanceItems)
+        await coordinator.writeService.persistBootstrap(progress: coordinator.progress, maintenanceItems: coordinator.maintenanceItems)
         let persistedAt = Date()
 
         if let loadedVideos = refreshState.videos {
-            videos = loadedVideos
+            coordinator.videos = loadedVideos
         }
         AppConsoleLogger.appLifecycle.debug(
             "refresh_ui_complete",
-            metadata: refreshUICompletionMetadata(.init(
+            metadata: support.refreshUICompletionMetadata(.init(
                 currentChannelID: currentChannelID,
                 includesVideos: includesVideos,
                 progress: nextProgress,
@@ -294,76 +208,12 @@ extension FeedCacheCoordinator {
         )
     }
 
-    func logRefreshUIStart(
-        currentChannelID: String?,
-        includesVideos: Bool
-    ) {
-        AppConsoleLogger.appLifecycle.debug(
-            "refresh_ui_start",
-            metadata: [
-                "current_channel": currentChannelID ?? "none",
-                "includes_videos": includesVideos ? "true" : "false",
-                "main_thread": AppConsoleLogger.mainThreadFlag()
-            ]
-        )
-    }
-
-    func applyRefreshUIState(
-        progress nextProgress: CacheProgress,
-        maintenanceItems nextMaintenanceItems: [ChannelMaintenanceItem],
-        currentChannelID: String?,
-        includesVideos: Bool
-    ) {
-        progress = nextProgress
-        maintenanceItems = nextMaintenanceItems
-        RuntimeDiagnostics.shared.record(
-            "refresh_ui_applied",
-            detail: "UI 状態を反映",
-            metadata: [
-                "currentChannelID": currentChannelID ?? "",
-                "maintenanceCount": String(nextMaintenanceItems.count),
-                "cachedVideos": String(nextProgress.cachedVideos),
-                "cachedChannels": String(nextProgress.cachedChannels),
-                "includesVideos": includesVideos ? "true" : "false"
-            ]
-        )
-    }
-
-    private struct RefreshUICompletionMetadataParams {
-        let currentChannelID: String?
-        let includesVideos: Bool
-        let progress: CacheProgress
-        let maintenanceItemCount: Int
-        let loadedVideoCount: Int?
-        let startedAt: Date
-        let snapshotLoadedAt: Date
-        let homeStatusUpdatedAt: Date
-        let persistedAt: Date
-    }
-
-    private func refreshUICompletionMetadata(_ params: RefreshUICompletionMetadataParams) -> [String: String] {
-        [
-            "current_channel": params.currentChannelID ?? "none",
-            "includes_videos": params.includesVideos ? "true" : "false",
-            "cached_channels": String(params.progress.cachedChannels),
-            "cached_videos": String(params.progress.cachedVideos),
-            "loaded_videos": params.loadedVideoCount.map(String.init) ?? "",
-            "maintenance_items": String(params.maintenanceItemCount),
-            "snapshot_ms": AppConsoleLogger.elapsedMilliseconds(from: params.startedAt, to: params.snapshotLoadedAt),
-            "home_status_ms": AppConsoleLogger.elapsedMilliseconds(from: params.snapshotLoadedAt, to: params.homeStatusUpdatedAt),
-            "persist_ms": AppConsoleLogger.elapsedMilliseconds(from: params.homeStatusUpdatedAt, to: params.persistedAt),
-            "videos_ms": params.includesVideos ? AppConsoleLogger.elapsedMilliseconds(from: params.persistedAt, to: Date()) : "0",
-            "elapsed_ms": AppConsoleLogger.elapsedMilliseconds(since: params.startedAt),
-            "main_thread": AppConsoleLogger.mainThreadFlag()
-        ]
-    }
-
     func completeImportedChannelUpdate(channels: [String], importedChannelIDs: [String]) async {
-        resetRemoteSearchSnapshotCache()
-        self.channels = channels
-        freshnessInterval = TimeInterval(max(channels.count, 1) * 60)
-        _ = await performConsistencyMaintenanceIfNeeded(force: true)
-        await bootstrapMaintenance()
+        coordinator.resetRemoteSearchSnapshotCache()
+        coordinator.channels = channels
+        coordinator.freshnessInterval = TimeInterval(max(channels.count, 1) * 60)
+        _ = await coordinator.performConsistencyMaintenanceIfNeeded(force: true)
+        await coordinator.bootstrapMaintenance()
 
         if !AppLaunchMode.current.usesMockData {
             scheduleImportedChannelRefresh(channelIDs: importedChannelIDs)
@@ -372,17 +222,17 @@ extension FeedCacheCoordinator {
 
     func scheduleImportedChannelRefresh(channelIDs: [String]) {
         guard !channelIDs.isEmpty else { return }
-        guard importRefreshTask == nil else { return }
+        guard coordinator.importRefreshTask == nil else { return }
 
-        importRefreshTask = Task {
+        coordinator.importRefreshTask = Task {
             await refreshImportedChannels(channelIDs)
-            importRefreshTask = nil
+            coordinator.importRefreshTask = nil
         }
     }
 
     func refreshImportedChannels(_ importedChannelIDs: [String]) async {
-        let snapshot = await readService.loadSnapshot()
-        let states = dictionaryKeepingLastValue(snapshot.channels.map { ($0.channelID, $0) })
+        let snapshot = await coordinator.readService.loadSnapshot()
+        let states = coordinator.dictionaryKeepingLastValue(snapshot.channels.map { ($0.channelID, $0) })
         let cachedVideoChannelIDs = Set(snapshot.videos.map(\.channelID))
         let prioritizedChannelIDs = importedChannelIDs.sorted { lhs, rhs in
             let lhsNeedsWarmup = states[lhs]?.channelTitle == nil || !cachedVideoChannelIDs.contains(lhs)
@@ -394,10 +244,10 @@ extension FeedCacheCoordinator {
         }
 
         for channelID in prioritizedChannelIDs {
-            _ = await self.processChannel(channelID, states: states)
+            _ = await coordinator.processChannel(channelID, states: states)
         }
 
-        _ = await performConsistencyMaintenanceIfNeeded(force: false)
-        await refreshUI(currentChannelID: nil, isRunning: false, lastError: progress.lastError)
+        _ = await coordinator.performConsistencyMaintenanceIfNeeded(force: false)
+        await refreshUI(currentChannelID: nil, isRunning: false, lastError: coordinator.progress.lastError)
     }
 }
